feat(dotfiles): improve PATH, MANPATH, INFOPATH handling
The `append_to`/`prepend_to` functions are dumb and repetitive, but they
get the job done. Now we correctly handle MANPATH (ie. not adding it
redundantly in subshells) and INFOPATH.

At least on macOS, we have these config files:

  /etc/manpaths

        /usr/share/man
        /usr/local/share/man

  /etc/manpaths.d/40-XQuartz (etc)

        /opt/X11/share/man

Even though they aren't in the MANPATH, they still get included in the
search path. ie. on opening a new shell, MANPATH will be something like
`/opt/homebrew/share/man` and all of these work:

    man screencapture # ie. from /usr/share/man
    man xmpwrite # ie. from /opt/X11/share/man
    man nvim # ie. from /opt/homebrew/share/man

I don't know if there is an equivalent to /etc/manpaths for INFOPATH,
but on opening a shell, INFOPATH starts off like:

    /opt/homebrew/share/info

And these work:

    man screen # ie. from /usr/share/info
    man gnupg # ie. from /opt/homebrew/share/info