refactor(shell): attempt arch detection
As noted inline, this doesn't actually work as expected due to
architecture preference inheritance. But I am adding it as a "refactor"
seeing as it edges in the direction we want to go, for educational
purposes.

I'll follow up presently with an approach that uses `/usr/bin/uname -m`
instead, which should work fine everywhere.