refactor(fig): use `uname` instead of `arch`
Well, I thought this was going to work, but it didn't... seems `uname`
inherits from the environment in almost exactly the same way `node` +
`arch` does.

ie.

output in shell: arm64
output in node: x86_64

You can see this in action here:

    $arch -x86_64 uname -m
    x86_64
    $ uname -m
    arm64

I say "almost" because with `arch` it actually when to "ia32".