fix(shell): detect arm64 correctly
So this is an enormous hack, but I don't have much choice.

If you run this in an x86_64 shell on an arm64 machine, it will end up
detecting your arch as arm64. Maybe that's not you want, but you should
probably be running an arm64 shell on an arm64 machine.

Homebrew just accepts whatever `uname -m` tells it. So that behaves
differently: if you're running in an x86_64 shell then it will assume
you're really on x86_64 and install into `/usr/local/bin/`.

This is a revert of 2290414ec1be78a3fdbc859f97127f324f0f5a60, with the
above changes squashed in.