chore(vim): remove references to Macvim
Haven't used it to do work in over 10 years. Have really only opened it
about once a year while troubleshooting rendering issues to answer the
question, "how would this render in Macvim?".

Well, in the update to Big Sur, the application when missing off the
disk, so I figure now is as good a time as any to stop pretending that I
am ever going to use this again. 😂