refactor(vim): remove `has('gui')` guard from wincent#functions#keynote()
Seems to work fine in Neovim in tmux in iTerm. For some reason, on my
second test of this, the text came out in black always, even when the
inspector said it shouldn't. But the exact same file shows text color
correctly just fine in Safari, so I am chalking this up to some bizarre
Chrome bug.