refactor(vim): tighten up temporary directory settings
No behavioral change expected here, but trying to stick more closely to
what the docs say. `'backupdir'` and `'undodir'` seem to be behaving as
though the "//" suffix was there all along.

Added the `file=NONE` suffix to `'viminfofile'`/`'shadafile'` because
that seems a more robust way of ensuring that we don't get any
root-owned files being written in user-owned locations.