refactor(vim)!: make ~/.vim/* symlink to ~/.config/nvim/*
As noted here:

    https://twitter.com/wincent/status/1401289744982675459

the symlink has been pointing in the opposite direction for just over 4
years now. It is time for a change.

This was originally motivated by a bug on my work machine where I had
the `$PATH` messed up intermittently, which meant that I would sometimes
open a file with Vim instead of Neovim without meaning to. Vim would
then complain about an undo file made by a newer version of Vim, or
something like that.

So I fixed the `$PATH` bug but in thinking about keeping the backup,
view, undo, and swap files separate between Neovim and Vim I started to
wonder whether inverting the symlink would be a good idea. It is churn,
but I think it is a good long-term ploy, because it opens up the
possibility of me diverging a little bit more in the future (eg. more
Lua-only stuff on the Neovim side, and letting the Vim config sit
as-is).

Highly possible that I've broken something here, so will follow up if i
find any problems.