refactor: make "main" the default branch
There are two kinds of breakage that I was worried about with the move
from "master" towards "main":

1. Tools which assume "master".
2. Existing links to "master"-located things.

So, because of "1" my plan was to wait a while before switching my
default. I was thinking of only starting new projects on the "main"
branch and not updating existing projects so as to avoid "2".

But I realized we can have our cake and eat it too by just making sure
that "master" is always in sync with "main". We do it with a hook that I
have tested lightly and hope works. 😂

In theory, this means that people who cloned or linked to "master" a
while ago will continue to have things working just fine. And new folks
will land on the repo page and see that the default is "main" and get
that.

Now, there are some places where I think I can't change this (gitweb,
for example; I don't know how to force that to show "main" as the
default). And there are a few type-"1" problems still noted inline in
comments here, like things that I want to update to be
default-branch-aware.

Closes: https://github.com/wincent/wincent/issues/100