feat: add post-receive alternative to our pre-push hook
This moves the responsibility for updating "master" to the server, which
means I don't have to set it up on multiple machines. I think I like
this better, and I have the luxury of being able to install this hook on
my own server (which syncs over to GitHub etc every hour).

Now, if I want to push a "hotfix" to GitHub directly, I can still do
that:

    git push github

But it will only push "main", and "master" will sync over within an
hour. I could reduce the interval, or make these forks push immediately
as a side-effect. Or I could just push both.

    git push github
    git push github main

The one downside here is that local "master" doesn't sync, so there
still may be some utility in have a pre-push hook that does it at that
time for you. I will probably do that in the next commit.