refactor: tweak pre-push to only sync the local branch
As we now have a post-receive hook installed on the remote server, the
pre-push is just a nice-to-have to keep the local branch up-to-date (and
really, pretty superfluous because I should never be checking it out,
touching it, or even looking at it). Basically just for my OCD, so that
I don't get annoyed by seeing the two branches not pointing at the same
commit.

Note that it only happens when I push, which seems like an acceptable
interval for it to happen at.