chore(ssh): add some more settings related to host keys
These are all a bit mutually overlapping and the exact effect of the
combination is confusing, but in my testing seems to work well.

I originally had:

- `NoHostAuthenticationForLocalhost yes` (disables host-based
  authentication[0])

In this commit I am somewhat redundantly adding:

- `GlobalKnownHostsFile /dev/null`
- `UpdateHostKeys no` (don't ask user about host keys or write them to
  file)
- `UserKnownHostsFile /dev/null`
- `StrictHostKeyChecking no` (automatically adds new keys and also
  connections to changed hosts to proceed)

If I understand correctly the combination of `StrictHostKeyChecking` and
`GlobalKnownHostsFile`/`UserKnownHostsFile` are critical: they say to
automatically accept/update any key, but because we write to /dev/null
nothing is actually persisted to disk. I also think a combo of
`UpdateHostKeys` and `StrictHostKeyChecking` might have done the trick
too.

[0]: https://en.wikibooks.org/wiki/OpenSSH/Cookbook/Host-based_Authentication