feat: add initial support for running on Debian/Codespaces
At this point it is just a stub that shows I can detect Debian as a
Linux variant and run a simple new aspect on it that installs two
packages with `apt-get`.

The idea is that you can override the "linux" platform settings by
providing something more specific, like "linux.debian", as shown here.
Note that that's _override_ and not _merge_. If you want to merge/reuse
you have to find other mechanisms, but Fig is TS, and aspects are TS, so
you can acheive reuse inside them by any means you want.

Note the type changes. I think it makes sense to continue defining
supported "platform" types, but I am in a bit of a hurry so I
hard-coded a conditional to check to see whether "linux.debian" should
apply. Ideally, I would be able to make it smart enough to do a template
literal:

```
`${platform}.${distribution}`
```

and have TS see whether that is a valid type or not. I'll look into that
later.