fix: address new TypeScript errors
Caused by the update from `typescript` 3.9.3 to 4.3.2 (and possibly also
the update from `@types/node` 14.0.5 to 15.12.2).

Some of these are unsavory, others no so much.

    error TS2790: The operand of a 'delete' operator must be optional.

    delete Context.currentAspect;

Fair enough. Good catch, even, I'd say.

    error TS2345: Argument of type '(value: unknown) => void' is not
    assignable to parameter of type '() => void'.

    cursorTo(process.stderr, 0, undefined, resolve);

Cool, cool. Easily fixed by adding a type parameter.

    error TS2532: Object is possibly 'undefined'.

    if (stat.isDirectory()) {

    error TS2532: Object is possibly 'undefined'.

    } else if (stat.isFile()) {

Sigh. It's because the `stat` call is overloaded, and TypeScript
picks the last one (ie. the most general one), which returns `Stats |
BigIntStats | undefined`. There is a more specific overload (the first
one), which returns `Stats` and applies when the options parameter is
`undefined`, which is our case here, but I don't know of any way to
force TS to "choose" that override or infer it as the right one. So, I
unhappily write out the return type by hand.

Similar thing here but with the promise-based API:

    error TS2322: Type '(path: PathLike) => Promise<Stats |
    BigIntStats>' is not assignable to type ...

    stat(

Similar woes here with an overloaded function, and similar workaround:
spell out the type instead of using the `ReturnType<>` utility.

    error TS2322: Type 'String' is not assignable to type 'Path'.
    Type 'String' is not assignable to type 'string'.

    return Object.defineProperties(new String(string), {

This is a sad one. I am abusing `new String()` here to make decorate
string objects, but TS makes that very hard as it is not a common idiom
(the problem is, I can't think of a better tool for the job for this
specific use case). I tried changing the type of `Path` to show that it
is a `String` and not a `string`, but that caused a cascade of over 100
errors, because all the call sites started complaining about the same
problem. This might be worth fixing and doing "right" at some point, but
it will take a lot of fiddly so I am going to punt on it for now.