fix(dotfiles): make "git zap" alias a bit friendlier
Still does the same thing, but with less irrelevant console spew
because it iterates over remote branches and syncs those instead of
iterating over local branches (which often won't have a corresponding
remote ref).

Note I also added a modeline because the way <Tab> wasn't producing hard
tabs in this file was driving me crazy. It's an ERB statement (ie. won't
produce output in the filled template) containing a JS comment (ie. so
it won't blow up when the template gets filled), but Vim still sees it
as a valid modeline.

I never bothered to implement ERB comment syntax (ie. `<%#`) because I
didn't have a need for it at the time, but maybe it is something I
should do. I expect it would be pretty easy to add.