refactor(dotfiles): use real ERB comments
Not really "real" because ERB comments don't allow you to slurp newlines
with `-%>`, even though they probably should. You see, on the one hand,
they kind of behave like `#` comment in Ruby (ie. they comment
everything to the end of the line, and they are followed by a newlines)
but on the other, they allow trailing text; eg:

   one<%# a comment %>            one
   two                            two

   one<%# a comment -%>           one
   two                            two

   one <%# a comment %>two        one two