feat: make Clipper work in Codespaces
Set up reverse port forward and tweak to get things working on Debian
with `socat`.

Note that this technically means we don't need `nc` installed any more,
but it doesn't hurt to have it (we're in a container; I am not too
worried about having random utils installed in it, and it may prove
useful for me to manually do things).