feat(dotfiles): add "git zip"
Counterpart to "git zap":

- "git zap" makes local branches look like remote ("zap" because its
  destructive)
- "git zip" archives away local branches that are no longer on the
  remote

Basically these are to support a workflow where:

- I do almost all real development on a remote server.
- I keep a local repo more or less in sync with that.
- I use the upstream origin as the conduit for the sync.
- The origin itself is configured to aggressively delete branches when
  they are closed/merged.

So, basically, you push to the origin regularly... you occasionally pull
stuff down to the local repo with `git zap` (and repeat whenever you
want to freshen the local copy)... and finally, when stuff on the origin
is "done" (merged or closed), you use `git zip` to zip (archive) it all
away.

Will wind up with a bunch of "attic/wincent/..." branches that I can
ignore (with grep or other patterns) and will probably end up making a
"git sweep" or something to sweep old stuff out of the attic from time
to time.