fix(zsh): make up/down do proper history search in Debian
Not sure why, but this wasn't working in Debian.

Debian has `/etc/zsh/zshrc` which defines things like:

    typeset -A key
    key=(
        Up         "${terminfo[kcuu1]}"
        Down       "${terminfo[kcud1]}"
        ...

macOS doesn't have that though. It does have a `$key` variable set
somewhere else (not sure where). Whatever the case, `$key[Up]` etc works
on Debian, but not on macOS. Haven't tried it on Arch yet.

Things that don't work on Debian:

    bindkey '^[[A' history-substring-search-up
    bindkey '^[[B' history-substring-search-down

    bindkey '\e[A' history-substring-search-up
    bindkey '\e[B' history-substring-search-down

    bindkey '^[OA' history-substring-search-up
    bindkey '^[OB' history-substring-search-down