fix(vim): allow `rollup --watch` to work by setting `backupcopy=yes`
This has bitten me before with respect to running `crontab -e` (see
911aba71ee8572f283b8), but yesterday was also hitting me because `rollup
--watch` was only picking up the first change to any given file.

The answer lies here:

    https://github.com/rollup/rollup/issues/1666#issuecomment-547613081

Rollup watches files, not directories, which means that when Neovim uses
its default `backupcopy=auto` strategy to update file contents on disk,
things go awry.

Specifically, from `:h backupcopy` we see this summary:

    The main values are:
    "yes"   make a copy of the file and overwrite the original one
    "no"    rename the file and write a new one
    "auto"  one of the previous, what works best

So, Vim's decision for "what works best" in this case is "no". IIRC,
when you save a file, it will try to rename it and then write a new copy
in its place; if the rename fails, it will fallback and try the other
strategy, but in practice, it's really always going to do the rename.

As explained in great in this Stack Exchange answer:

    https://vi.stackexchange.com/a/25040

Rollup ends up watching the renamed file (because it is watching by
inode). You save it again, and the renamed file gets deleted to make way
for the new backup, so Rollup sees the file go away and never watches
anything related again.

The fix is to force "yes" mode, which as described here:

    https://vi.stackexchange.com/a/11649

is going create a backup copy of the file being saved and then directly
overwrite the original. So, Rollup's watching is going to work fine with
that; no inodes were harmed in the process. I guess the reason why Vim
doesn't pick this strategy as the default is because it is doing more
work to save the file, so it might be slower for big files. But I would
rather have slow-and-working than fast-and-broken any day!

I also think that with version control, it is probably pointless to
write these backups anyway. I don't recall ever having to recover
anything at all from a Vim backup file in well over a decade of use...

At the moment I have `nobackup` and `writebackup` set (ie. don't keep
backups lying around, but do make them temporarily during writes).
That's the default, but I'm not sure it's all that useful...