fix(vim): try to handle more focus/blur edge cases
Probably not a definitive fix yet, and I don't have time to keep testing
this right now, but in an initial spot-check it seems to be a promising
approach. Instead of trying to blur/focus once only, we let Vim fire all
the autocommands it wants and we potentially focus multiple times. But
we do so with much more granular whitelists/blacklists that are more
specific to the concrete things that we want to turn on or off, or save
and restore etc.

See: https://github.com/wincent/wincent/issues/106