feat: turn off 'relativenumber' in blurred windows
Almost by definition, relative numbers are only really useful in the
currently focused window (ie. so you can jump there or do something with
a motion). In blurred windows, having real numbers tends to be much more
useful (ie. you can at a glance see where any given line is relative to
the start of the buffer).

So here's a bit of an experimental take on updating this dynamically. As
you can see, some hack to accommodate Goyo is needed because it creates
a number of "invisible" windows at top/bottom/left/right as spacers and
if we're not careful we end up ruining the entire effect by showing
numbers in those.