refactor(vim): port init.vim to init.lua
To celebrate the release of Neovim v0.5.0.

This is a quick and dirty port (ie. not attempting to use "idiomatic"
Lua constructs).

We basically fork the old "init.vim" over to ".vim/vimrc" and rewrite
its contents into Lua at "~/.config/nvim/init.lua". This is just a
1-to-1 translation, with the `if ('nvim')` checks removed from both
sides, keeping/discarding the appropriate side of the conditional block
depending on whether we're in Vim or Neovim.