refactor(vim): move nvim-compe set-up into separate file
This has to happen after `packadd`, so let's just get it out of init.lua
entirely.

Note the use of `pcall`, so that if we do disable this by commenting out
the `packadd`, we don't get a spew of red error text. If you saw this
error (unlikely) and wanted to troubleshoot it, you could just come in
here and comment out the `pcall`. I am not even sure if this is worth
keeping in here; maybe it is better to just fail loudly when things go
wrong 🤷.