feat(dotfiles): make tmux window labels a bit more stable
1. Always the same width, regardless of whether they have a flag or not.
2. With at least one space on each side.

eg.

    $left  1:vim*   2:zsh-   3:zsh    4:zsh    5:zsh    6:zsh
          aaaaaaaa bbbbbbbb cccccccc dddddddd eeeeeeee ffffffff

In the example above, "aaaaaaaa" is the current window (indicated by the
"*" flag), "bbbbbbbb" is the last window we were in (indicated by the
"-" flag), and all the others ("cccccccc", "dddddddd" etc) are other
windows. Note that the current window (1) has a background red highlight
that extends 1 character beyond the boundary of the visible text (eg.
there are 8 cells with a red background, vs the visible text itself --
"1:vim*" -- which is 6 long).

This compares with the old style, which looked like this:

    $left 1:vim* 2:zsh- 3:zsh 4:zsh 5:zsh 6:zsh
          aaaaaa bbbbbb ccccc ddddd eeeee fffff

So, that was more compact, but I never run out of space in these things
anyway. I usually have about 2 or 3 windows open, and rarely more than
6. Very rarely more than 8.