refactor(vim): move settings into init.lua
So I recorded a screencast about this in which I discussed the load
order issues here. In the old set-up, `:scriptnames` shows basically
this:

  1. init.vim
  ...
  2. settings.vim
  3. statusline.vim
  ...
  4. after/*.vim

After porting to Lua we don't see:

  1. init.vim
  ...
  2. settings.lua
  3. statusline.vim
  ...
  4. after/*.vim

but rather:

  1. init.vim
  ...
  2. statusline.vim
  ...
  3. settings.lua
  4. after/*.vim

ie. the Lua files get sourced way later, right at the end.

Given that options can affect the behavior of pretty much anything that
happens after them, seems like a good idea to set options early on. And
what better place for that than the new "vimrc" (ie. "init.lua")?

Seeing as I am dumping a bunch of crap into this file that I previously
tried to keep rather empty, I am introducing manual folding with section
headers as an organization mechanism.

Additional changes to note during the migration:

- Simplified `os.getenv('BLAH')` to `vim.env.BLAH`.
- Refactored "is root" check to use `vim.env.USER == 'root'` instead of
  the indirect `SUDO_USER` test. Note that this _is_ going to have
  ramifications for my user on Codespaces, where you run as root in the
  VM; it means that I won't have undofiles, view files etc. I'm not
  convinced that this is the right thing, so I might end up changing it;
  most likely, on the Codespaces host I'll probably add an override that
  re-sets those options to the desired values.
- Added a number of `vim.fn.expand()` calls because many options weren't
  working properly with "~" (eg. `'undodir'` and friends).
- Simplified a `vim.opt.loadplugins:get()` read access to just
  `vim.o.loadplugins`.
- Fix bug where I was eagerly calling `wincent#settings#foldtext()`
  instead of passing it as a string representing a function to be called
  later (ie. in the `'foldtext'` option).

And I think that about covers it.