feat(vim): add :Glow command for Glow previews
There is also:

    https://github.com/npxbr/glow.nvim

but:

1. I don't want another plug-in.
2. That one is small, but still bigger than I want/need (eg. it has code
   for installing Glow).
3. The floating window doesn't play nicely with my focus styling.
4. Seems to be lacking a `:startinsert` call that would make it
   actually useful (without it, you just see the top of the file and
   can't page through it).

Dispatch gives us what I want though: a way to conveniently jump to
another tmux window, preview the file and then "q" to get back on with
my life. No attempt at any fancy validation here; for example, if you
run `:Glow` on a Vimscript file, it will happily render it in Glow, and
that seems perfectly fine tbh.

Not super useful/pleasant outside of tmux (for example, on macOS, it
runs `glow` in an iTerm tab, but doesn't switch focus to it), but that's
ok too; I'm always in tmux.

Alternatives considered: using `:terminal` or `termopen()`.