perf(vim): shave about 20ms off startup time
By avoiding eager-loading of an expensive Ferret file. We can afford to
set `g:FerretExecutableArguments` lazily because they are only consulted
on the first actual search.

Original I was going to be "clever" and do something like this instead:

    autocmd! SourcePost */ferret/autoload/ferret/private.vim ++once :let g:FerretExecutableArguments={
            \   'rg': ferret#get_default_arguments('rg') . ' --hidden --glob !.git'
            \ }

but it's really not that clever. If we do `SourcePost`, we end up
setting the arguments just a wee bit too late. There is a loop inside
`ferret#private#executable` and if you add an `echo` in there you can
see that on the first time through the global isn't set yet, but it is
on the 2nd and 2rd time.

Likewise, `SourcePre` is flawed because using it will cause
"private.vim" to be loaded twice: once because something calls it, and
once because our `SourcePre` call ends up loading "private.vim" as a
side-effect.

Overall, I think the `ferret#get_default_arguments()` design is flawed
because there is no way to be fast and simultaneously avoid the
chicken-and-egg situation described above. We just side-step the issue
by hard-coding the darn string in here, and assuming we have a recent
`rg` on the system, which we usually do. If `rg` breaks its API in the
future, I'll have to rethink this, but for now I am sticking to the
low-tech workaround.