perf(vim): avoid expensive projectionist detection
Was rather horrified to see vim-projectionist adding over 80ms to Vim
startup time due to these heuristics, when opening an empty buffer. With
the changes in this commit, it now takes about 17ms. When opening a
non-empty buffer, it's basically free.

So, what was going on here? Well, my "catch-all" pattern of "*" was
wreaking a bit of havoc, especially when combined with my
combinatorially exploding ".js"/".jsx"/".ts"/".tsx" config.
Projectionist would create "roots" all the way up from the file being
edited right up to the filesystem root ("/").

Scoping filetype-specific config to concrete root markers (eg.
"package.json" for ".js"/".jsx", "tsconfig.json" for ".ts"/".tsx",
"bsconfig.json" for ".re" etc) makes a big difference, cutting off about
half the time.

I tried changing "*" to "." for my catch all but that was no good. It
basically fixed the issue of the hierarchy of "roots", but it meant that
if I tried to open, say, a "*.c" file in a sibling directory then I
wouldn't get alternates.

The behavior of "*", on the other hand, is quite weird. Opening a "*.c"
file in a sibling ends up creating multiple roots for that too, but not
all the way to the root; for example, in my dotfiles repo, a sibling of
"wikitext", I could open "../wikitext/src/ext/wikitext/ary.c", and
Projectionist would set up roots for that buffer at each of the
following:

- "$path_to/wikitext"
- "$path_to/wikitext/src"
- "$path_to/wikitext/src/ext"
- "$path_to/wikitext/src/ext/wikitext"

All of this determined by looking at the value of `b:projectionist`.

Whatever, it basically means that the "project" in "projectionist"
is deadly serious. It isn't supposed to work properly outside of a
"project". It is not literally a replacement for the venerable a.vim:

    https://www.vim.org/scripts/script.php?script_id=31

It assumes/requires a bit more structure.  So, I ruefully declared that
`Makefile` should be present to signify a C project. This will probably
be ok.

In the future, I might consider trying to do some lazy detection with:

    autocmd User ProjectionistDetect call wincent#my_magical_detection_thing()

but I am a bit dubious of it; a search of all code on GitHub turns up
12(!) hits of usages, about 3 of which look to be mindless copies,
several belong to tpope himself, and that leaves probably one other
human being in the universe that knows how/what to pass to
`projectionist#append()`... But maybe I'll think about it some day. It
might be as easy as, if the current file is a ".c" or ".h" file, append
a mapping for it in its directory etc. Will cross that bridge if and
when we come to it.