refactor(vim): hard fork pretty much all settings
All the things that were symlinks into "~/.config/nvim" have been copied
into "~/.vim" so that they can diverge freely. The only things that are
still shared are:

- "~/.vim/pack" -> "~/.config/nvim/pack" (ie. all the submodules).
- "~/.vim/ultisnips" -> "~/.config/nvim/ultisnips"
- "~/.vim/ultisnips-private" -> "~/.config/nvim/ultisnips-private"

Now, I may one day allow even the "pack" stuff to diverge but for now
sharing it is harmless because stuff is only pulled in from there by
explicit ":packadd" commands.

Likewise, UltiSnips might diverge if I switch to a more Neovim-ish
snippet solution in the future (probably something with tighter
integration into nvim-compe, and probably written in Lua instead of
Python), but I am not in an immediate hurry to do that yet[*].

[*]: Normally when I say this you see a commit enter my dotfiles repo
with the corresponding change only a few hours later 😆.