feat(vim): start migrating from UltiSnips to LuaSnip
This is still very incomplete and a work-in-progress, but I wanted to
show my work as I go. This clearly shows promise, because LuaSnip is
ridiculously powerful. I only have some simple demo snippets at this
point, but will start fleshing out the more important ones in the coming
days.

Getting the <Tab> and <BS> nuances right is very tricky, so expect this
to change as I play with it and start seeing how more complex snippets
interact with completion.