fix(vim): make <Tab> work at the end of the line
This didn't used to be necessary in my old set-up, but then again, my
old set-up had a different set of quirks. I don't like the asynchrony
here, but I don't think there is anything I can do about it because
that's just the way `compe#complete()` is (feature, not bug). In my
brief testing, this seems to work well enough.