refactor(vim): add some utility methods
Just playing with module-loading patterns. I don't think I'll actually
use any of these. Vanilla Lua is probably as good as anything:

    -- has()
    if has('compe') then
      compe = require'compe'
      -- do stuff with compe...
    end

    -- safe_require()
    compe = safe_require('compe')
    if compe then
      -- do stuff with compe...
    end

    -- using() (could also have called this import())
    using('compe', function(compe)
      -- do stuff with compe...
    end)

    -- vanilla Lua
    local has_compe, compe = pcall(require, 'compe')
    if has_compe then
      -- do stuff with compe...
    end

Of all of these, `using()` is probably the worst, because if you depend
on two modules that might not be there (ie. due to not having been
`packadd`-ed) then you have to do:

    using('compe', function(compe)
      using('luasnip', function(luasnip)
        -- do stuff with compe and luasnip
      end)
    end)

That suggests an obvious extension to the API:

    using('compe', 'luasnip', function(compe, luasnip)
      -- do stuff with compe and luasnip
    end)

or similar.