refactor: extract map functions out into modules
This shows the very granular direction I am trying to head in. It's true
that by the time Neovim finishes booting, almost everything in here is
going to be evaluated, but in case I turn stuff off, I want to be sure
that only the stuff that is actually needed gets pulled in.