refactor(vim): remove unused functions
So I had these in there just as examples and wasn't convinced that I
should/would use them. I think the last commit shows that it is totally
reasonable to do the kind of thing (conditional `require`) that these
modules seek to address in vanilla Lua, so out they go.

On the bright side, the Git history now has an example of how to use
`unpack()` to forward varags to other functions. 😀