feat(vim): add "require" snippet
First cut of one, anyway. In some ways this works way better than
solutions I've had in the past, but I am still not 100% happy with it.

The good: <Tab> and <S-Tab> work great for moving back and forth (in the
past I had trouble moving back); note that I had to change some
"noremap" mappings to get that working.

The bad: Dynamic update of filename based on local binding name only
happens on leaving the node (ie. hitting <Tab>) instead of during
typing, like it used to in UltiSnips.

The ugly: If I <S-Tab> back, having edited the filename, then I lose the
filename edits if I edit the local binding name.

I suspect there is probably a solution to either of both of my concerns,
so I'll keep looking.