refactor(vim): add augroup() and autocmd() for simple use cases
This will evidently grow as I port more things to use it, and it is very
primitive. For example, you can clear an autocmd with Vim commands and
we'll never garbage collect the contents of
`wincent.g.autocommand_callbacks`. But that seems fine; in general, the
kinds of things I am going to use this for are expected to live the
lifetime of the editor session (later on, if I add support for
buffer-local autocmds or the `++once` option, then I'll have to
implement some kind of clean-up).