refactor(vim): audit use of map vs noremap variants
All the things that may return something special, like `<Plug>whatever`
use a "map" variant. The others can use "noremap" variants, as they
return basic keystrokes for the most part, like "<C-n>" and so on.

The one exception is `smart_bs`, which may return something more
complicated:

    <C-\><C-o>:set expandtab<CR><BS><C-\><C-o>:set noexpandtab<CR>

but even that is all just primitive Vim stuff, so that works too.
(Tested by turning on `noexpandtab` and backspacing in trailing
whitespace).