refactor(nvim): port variables to Lua
I like getting rid of my `defer#defer` hack, replacing it with the
native `defer_fn()`.

On the other hand, I kind of hate the file naming here:

- aspects/nvim/files/.config/nvim/plugin/variables.lua
  (loaded eagerly during boot)
- aspects/nvim/files/.config/nvim/lua/wincent/variables/init.lua
  (loaded after start-up)

I could have called the second one "variables.lua" too, but that is just
as bad/confusing. I'll see if I can think of a better name tomorrow
after sleeping on it.