refactor(nvim): port foldtext() to Lua
Seeing as I want to use this as `v:lua.wincent.foldtext()`, that
requires me to re-jig my global handling a little bit. I think it makes
it simpler if I just assume that `require'wincent'` is going to be the
first thing in my top-level "init.lua"; that way, I can drop a bunch of
`require` statements from all over the place (should look cleaner), and
trust that the lazy-loading mechanism won't pull in anything that is not
actually needed.

I don't even think it conflicts with separate plugins I've made, like
Pinnacle (which you access via `require'wincent.pinnacle'.dump()` and so
forth.