refactor(lua): port lazy plugin-loading to Lua
Along the way, discovered that a 2018 commit in undotree
(a80159c9f5c238) probably broke the old code anyway, by renaming a
keymap target.

Includes "dispose" functionality that I thought I was going to need. In
the end, I don't, but it seems useful enough to leave in there anyway.