feat(nvim): use nvim-tree.lua to get tree views back again
nvim-tree.lua isn't a replacement for vim-dirvish (it can't do the "full
screen" thing yet), so I am keeping both.

One annoying bug to live with for now (autocloses as soon as you open a
file):

    https://github.com/kyazdani42/nvim-tree.lua/issues/502

The "full screen" thing may happen in the future, based on:

    https://github.com/kyazdani42/nvim-tree.lua/pull/288

but looking at the implementation I am not sure (see the use of `:only`,
which will definitely break the use case).

Rejigged mappings along the way so I can toggle this thing open with
`<LocalLeader>t` (mnemonic: [t]ree) and do the useful
`:NvimTreeFindFile` with `<LocalLeader>f` (mnemonic: [f]ind).

Giving up on fern.vim as I don't think it will ever be fast enough to
displace vim-dirvish. nvim-tree.lua is pretty fast; not quite as fast as
vim-dirvish, but totally adequate.