feat(nvim): make :W work in Neovim
In porting vim-eunuch.vim to vim-eunuch.lua, I was reminded that the old
trick doesn't work in Neovim. I had barely noticed because the truth is
that I had trained myself to _not_ open root-owned files for writing
from an unprivileged user.

But anyway, we can fix it. I'm doing this "inside" Vim because I want to
have a way of getting the password to `sudo` without having to find a
variant of `SUDO_ASKPASS` that works for each platform I use.

`:W` prompts for password, writes the file, caches the pass for 5
minutes.

`:W!` will prompt you for a password even if one is already cached
(useful if you entered a wrong pass and it got cached).

Security disclaimer: this asks for your password and stores it in a
variable for 5 minutes. Even after the variable contents are
garbage-collected, the text of the password may be recoverable from
memory using various tricks (debuggers etc). The password is
additionally (briefly) written to disk, where it is only as secure as
filesystem permissions allow (eg. does your OS return a truly private
`tempname()`? is the 700 mode adequately enforced on your filesystem?
are there other privileged users running on the system that would try to
read the pass? once the file is unlinked, when do the actual contents
get truly purged from storage? etc).

As I use this rarely, I am fine with all these trade-offs. But your
mileage may vary.

The other gotcha is that if you enter a wrong pass, Vim will pause while
`sudo` asks two more times for the (same) wrong password. At this point
if you want to retry you should use `:W!`. Again, I'm fine with this
trade-off because I don't think there is an easy workaround.