feat(dotfiles): teach `git root` to print root too
And show how it can be used.

Specifically, we want to make sure `safe-ruby` is in the `$PATH` if
possible.

Sure, I could also have just done a `git root` at the top to `cd` at the
start, but I didn't think of that until too late. In any case, I have a
preference for avoiding `cd` in shell scripts if possible because it can
be hostile to refactoring (ie. any change in order risks breaking things
because lines no longer run from the context they think they're running
from).