feat(nvim): make help available for lazy-loaded plugins
To understand what's going on here, you need to know:

1. `packadd!` adds to the 'runtimepath' but it doesn't load anything
2. `packadd` adds to the 'runtimepath' _and_ loads the plugin eagerly
3. As part of the |load-plugins| process, Vim will load all plugins that
   are found in the 'runtimepath'.

So, the net effect of this commit is:

1. `packadd!` in `init.lua` sets up the runtimepath.
2. |load-plugins| happens, so plugins get loaded.
3. |VimEnter| happens, so additional `packadd!` calls are made; this
   adds the plugins to the 'runtimepath', which means we can find their
   help, but the plugins aren't loaded yet.
4. Due to user interaction and `wincent.lazy` mapping/command stubs, the
   lazy plugins may be loaded for real later on, on demand.

Note that for lazy-loaded plugins, we call `packadd`, which will no-op
re-add them to the `runtimepath` again, and then load them, and then we
run the desired mapping/command.

The main gap in this abstraction: once a plugin like "Goyo" is in the
'runtimepath' but not loaded yet, you can invoke an autoloaded function
and that will cause the autoloaded function's file to get loaded. It
_doesn't_ load the rest of the plugin though. ie. I can run `:call
goyo#execute('!', '') and it will fire up Goyo, but it won't source
"plugin/goyo.vim" (which would define the real `:Goyo` command).

All of this suggests that we may want to roll this into a slightly
higher level abstraction to stop it leaking. But I don't think we'll
ever be able to plug the autoload hole.